Created model package and moved model objects into it.
