+++ src/model/Board.java
@@ -1,3 +1,4 @@
+package model;
 
 import java.util.ArrayList;
 import java.util.List;
