+++ src/model/MoveInterface.java
@@ -1,3 +1,4 @@
+package model;
 
 public interface MoveInterface {
