+++ src/model/MoveValidator.java
@@ -1,3 +1,4 @@
+package model;
 import java.util.ArrayList;
 
 public class MoveValidator {
