+++ src/model/PieceColor.java
@@ -1,3 +1,4 @@
+package model;
 
 public enum PieceColor {
     WHITE, BLACK
